ines.cpp: Don't randomize the battery-backed portion of CHR-RAM
The code to save the battery-backed portion of CHR-RAM off to the .sav
file was working, but the reload of the data wasn't working quite right.
There are actually 2 places where the entire VROM buffer was being
randomized, causing any data loaded from disk to get clobbered.

Tweak the math here to skip over the battery-backed portion when
randomizing the VROM buffer.